+++ cmd/rs.go
@@ -131,6 +131,7 @@ var (
 	endDate    string
 	maxRetry   int
 	finalKey   string
+	appendMode bool
 )
 
 func init() {
@@ -144,10 +145,11 @@ func init() {
 	lsBucketCmd2.Flags().StringVarP(&listMarker, "marker", "m", "", "list marker")
 	lsBucketCmd2.Flags().StringVarP(&prefix, "prefix", "p", "", "list by prefix")
 	lsBucketCmd2.Flags().StringVarP(&suffixes, "suffixes", "q", "", "list by key suffixes, separated by comma")
-	lsBucketCmd2.Flags().IntVarP(&maxRetry, "max-retry", "x", 20, "max retries when error occurred")
+	lsBucketCmd2.Flags().IntVarP(&maxRetry, "max-retry", "x", -1, "max retries when error occurred")
 	lsBucketCmd2.Flags().StringVarP(&outFile, "out", "o", "", "output file")
 	lsBucketCmd2.Flags().StringVarP(&startDate, "start", "s", "", "start date with format yyyy-mm-dd-hh-MM-ss")
 	lsBucketCmd2.Flags().StringVarP(&endDate, "end", "e", "", "end date with format yyyy-mm-dd-hh-MM-ss")
+	lsBucketCmd2.Flags().BoolVarP(&appendMode, "append", "a", false, "append to file")
 
 	moveCmd.Flags().BoolVarP(&mOverwrite, "overwrite", "w", false, "overwrite mode")
 	moveCmd.Flags().StringVarP(&finalKey, "key", "k", "", "filename saved in bucket")
@@ -163,9 +165,8 @@ func init() {
 func DirCache(cmd *cobra.Command, params []string) {
 	var cacheResultFile string
 	cacheRootPath := params[0]
-	if len(params) == 2 {
-		cacheResultFile = params[1]
-	}
+
+	cacheResultFile = outFile
 	if cacheResultFile == "" {
 		cacheResultFile = "stdout"
 	}
@@ -176,10 +177,6 @@ func DirCache(cmd *cobra.Command, params []string) {
 }
 
 func ListBucket2(cmd *cobra.Command, params []string) {
-	if maxRetry <= 0 {
-		fmt.Fprintf(os.Stderr, "maxRetry must be greater than 0\n")
-		os.Exit(1)
-	}
 	bucket := params[0]
 
 	var dateParser = func(datestr string) (time.Time, error) {
@@ -221,7 +218,7 @@ func ListBucket2(cmd *cobra.Command, params []string) {
 		}
 	}
 	bm := iqshell.GetBucketManager()
-	retErr := bm.ListBucket2(bucket, prefix, listMarker, outFile, "", start, end, sf, maxRetry)
+	retErr := bm.ListBucket2(bucket, prefix, listMarker, outFile, "", start, end, sf, maxRetry, appendMode)
 	if retErr != nil {
 		os.Exit(iqshell.STATUS_ERROR)
 	}