+++ iqshell/account.go
@@ -99,7 +99,7 @@ func DecryptSecretKey(accessKey, encryptedKey string) (string, error) {
 	return secretKey, nil
 }
 
-func setdb(acc Account) (err error) {
+func setdb(acc Account, accountOver bool) (err error) {
 	accDbPath := AccDBPath()
 	if accDbPath == "" {
 		return fmt.Errorf("empty account db path")
@@ -110,6 +110,19 @@ func setdb(acc Account) (err error) {
 		os.Exit(STATUS_HALT)
 	}
 	defer ldb.Close()
+
+	if !accountOver {
+
+		exists, hErr := ldb.Has([]byte(acc.Name), nil)
+		if hErr != nil {
+			err = hErr
+			return
+		}
+		if exists {
+			err = fmt.Errorf("Account Name: %s already exist in local db", acc.Name)
+			return
+		}
+	}
 
 	ldbWOpt := opt.WriteOptions{
 		Sync: true,
@@ -127,7 +140,7 @@ func setdb(acc Account) (err error) {
 	return
 }
 
-func SetAccount2(accessKey, secretKey, name, accPath, oldPath string) (err error) {
+func SetAccount2(accessKey, secretKey, name, accPath, oldPath string, accountOver bool) (err error) {
 	acc := Account{
 		Name:      name,
 		AccessKey: accessKey,
@@ -139,7 +152,7 @@ func SetAccount2(accessKey, secretKey, name, accPath, oldPath string) (err error
 		return
 	}
 
-	err = setdb(acc)
+	err = setdb(acc, accountOver)
 
 	return
 }
@@ -151,21 +164,21 @@ func SetAccount(acc Account, accPath, oldPath string) (err error) {
 	}
 	if _, sErr := os.Stat(QShellRootPath); sErr != nil {
 		if mErr := os.MkdirAll(QShellRootPath, 0755); mErr != nil {
-			err = fmt.Errorf("Mkdir `%s` error, %s", QShellRootPath, mErr)
+			err = fmt.Errorf("Mkdir `%s` error: %s", QShellRootPath, mErr)
 			return
 		}
 	}
 
 	accountFh, openErr := os.OpenFile(accPath, os.O_CREATE|os.O_RDWR, 0600)
 	if openErr != nil {
-		err = fmt.Errorf("Open account file error, %s", openErr)
+		err = fmt.Errorf("Open account file error: %s", openErr)
 		return
 	}
 	defer accountFh.Close()
 
 	oldAccountFh, openErr := os.OpenFile(oldPath, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, 0600)
 	if openErr != nil {
-		err = fmt.Errorf("Open account file error, %s", openErr)
+		err = fmt.Errorf("Open account file error: %s", openErr)
 		return
 	}
 	defer oldAccountFh.Close()